Move failsafe to util
