+++ src/utils/util.js
@@ -28,6 +28,8 @@ const axios = require('axios');
 const qs = require('qs');
 const https = require('https');
 const config = require(process.cwd() + '/config.json');
+const controlPlaneUrl = config.controlPlane.url;
+const sequelize = require("../db/sequelize");
 const { body, param, query } = require('express-validator');
 const { Sequelize } = require('sequelize');
 const apiDao = require('../dao/apiMetadata');
@@ -721,11 +723,11 @@ function appendAPIImageURL(subList, req, orgID) {
     });
 }
 
-async function appendSubscriptionPlanDetails(orgID, subscriptionPolicies) {
+async function appendSubscriptionPlanDetails(req, orgID, subscriptionPolicies) {
     let subscriptionPlans = [];
     if (subscriptionPolicies) {
         for (const policy of subscriptionPolicies) {
-            const subscriptionPlan = await loadSubscriptionPlan(orgID, policy.policyName);
+            const subscriptionPlan = await loadSubscriptionPlan(req, orgID, policy.policyName);
             subscriptionPlans.push({
                 policyID: subscriptionPlan.policyID,
                 displayName: subscriptionPlan.displayName,
@@ -739,10 +741,10 @@ async function appendSubscriptionPlanDetails(orgID, subscriptionPolicies) {
     return subscriptionPlans;
 }
 
-const loadSubscriptionPlan = async (orgID, policyName) => {
+const loadSubscriptionPlan = async (req, orgID, policyName) => {
 
     try {
-        const policyData = await apiDao.getSubscriptionPolicyByName(orgID, policyName);
+        const policyData = await getSubPolicyByName(req, orgID, policyName);
         if (policyData) {
             return new subscriptionPolicyDTO(policyData);
         } else {
@@ -812,6 +814,78 @@ async function listFiles(path) {
     return files;
 }
 
+const createSubscriptionPolicyFromListByName = async (subscriptionPolicies, policyName) => {
+    let createdPolicy;
+    await sequelize.transaction(async (t) => {
+        for (const policy of subscriptionPolicies) {
+            if (policy.policyName === policyName) {
+                const created = await apiDao.createSubscriptionPolicy(orgId, policy, t);
+                if (!created) {
+                    throw new CustomError(
+                        500,
+                        constants.ERROR_CODE[500],
+                        `Failed to create policy: ${policy.policyName || "unknown"}`
+                    );
+                }
+                createdPolicy = new subscriptionPolicyDTO(created);
+                break;
+            }
+        }
+    });
+    return createdPolicy;
+};
+
+// ***** Get CP Subscription Policies for Organization *****
+
+const getCPSubPolicies = async (req, orgIdentifier) => {
+    const subPoliciesUrl = `${controlPlaneUrl}/throttling-policies/subscription?organizationId=${orgIdentifier}`
+    const responseData = await invokeApiRequest(req, 'GET', subPoliciesUrl, null, null);
+    if (!responseData?.list || !Array.isArray(responseData.list)) {
+        throw new Error("Invalid subscription policies response format");
+    }
+    return responseData.list.map(item => ({
+        policyName: item.name,
+        displayName: item.displayName,
+        billingPlan: item.tierPlan,
+        description: item.description,
+        type: item.quotaPolicyType,
+        timeUnit: item.timeUnit,
+        unitTime: item.unitTime,
+        requestCount: item.requestCount,
+        dataAmount: null,
+        dataUnit: item.dataUnit,
+        EventCount: null
+    }));   
+}
+
+// ***** getSubscriptionPolicyByName Interceptor *****
+
+const getSubPolicyByName = async (req, orgId, policyName) => {
+    try {
+        let subscriptionPolicy = await apiDao.getSubscriptionPolicyByName(orgId, policyName);
+        if (!subscriptionPolicy) {
+            // Since sub-policy cannot be found, check CP for the sub-policy
+            const { ORGANIZATION_IDENTIFIER: orgIdentifier } = await adminDao.getOrganization(orgId);
+            const cPSubPolicies = await getCPSubPolicies(req, orgIdentifier);
+            if (!cPSubPolicies) {
+                // Sub-policies not available in CP for the org
+                return null;
+            }
+            const response = await createSubscriptionPolicyFromListByName(cPSubPolicies, policyName);
+            if (!response) {
+                // Matching sub-policy (name) was not found from CP to add
+                return null;
+            }
+            // Again get sub-policy since matching Sub-policy was added from CP
+            subscriptionPolicy = await apiDao.getSubscriptionPolicyByName(orgId, policyName);       
+        } 
+        return subscriptionPolicy;
+    } catch (error) {
+        console.error("Error occurred while getting Subscription Policy by name", error);
+        handleError(res, error);
+    }
+}
+
 module.exports = {
     loadMarkdown,
     renderTemplate,
@@ -839,5 +913,6 @@ module.exports = {
     tokenExchanger,
     listFiles,
     readDocFiles,
-    unzipDirectory
+    unzipDirectory,
+    getSubPolicyByName
 }
